+++ ks43team01/src/main/resources/mapper/UserMapper.xml
@@ -89,7 +89,7 @@
 	  		<result column="level_up_buyer_standard_reg_date" property="levelUpBuyerStandardRegDate"/>
 		</resultMap>
   	  
-  	  <update id="modifyBusiness" parameterType="SellerBusiness">
+  	<update id="modifyBusiness" parameterType="SellerBusiness">
   		/*판매자 비지니스수정처리*/  		
   		UPDATE
   			tb_u_expert_business_filed
@@ -104,7 +104,7 @@
 		
 	</update>
   	  
-  	  <update id="modifyUser" parameterType="User">
+	<update id="modifyUser" parameterType="User">
   	  	/*유저정보 수정*/
   	  	UPDATE
 	 		tb_u_all_user AS u
@@ -117,7 +117,7 @@
 			u.user_id_code =#{userIdCode};
   	  </update>
   	  	 	
-  	  <update id="modifyUserLevel" parameterType="User">
+  	<update id="modifyUserLevel" parameterType="User">
   		/*판매자 정보 입력시에 등급 3에서 2로 업데이트!!!*/
   		UPDATE
 			tb_u_all_user AS u
@@ -128,7 +128,7 @@
 			u.user_id_code = #{userIdCode};
   	  </update>
   	  
-  	  <insert id="addUserLevelExp" parameterType="UserLevelExp">
+  	<insert id="addUserLevelExp" parameterType="UserLevelExp">
 		/*로그인시에 구매자회원 경험치 누적!!!*/
 		INSERT INTO tb_l_level_buyer_standard_details 
 		(  level_up_buyer_standard_details_code
@@ -321,7 +321,7 @@
   	<select id="getAdminUserList" resultType="User">
   		/*회원모두보여주기*/
   		SELECT
-			user_id_code as userId
+			user_id_code as userIdCode
 			,user_pw as userPw
 			,user_name as userName
 			,user_contact as userContact
@@ -329,6 +329,8 @@
 			,user_withdraw_status as userStatus
 			,user_activity_area as userArea
 			,user_retention_Points as userPoint
+			,user_rating_buyer
+			,user_rating_seller
 			,user_rating as userRating
 			,user_reg_date as userRreDate
 			,user_update_date as userUpDate
