+++ ks43team01/src/main/resources/templates/adminpage/user/userList.html
@@ -22,7 +22,7 @@
                   <div class="x_content">
                       <div class="row">
                           <div class="col-sm-12">
-                            <div class="card-box table-responsive" style="overflow: auto">
+                            <div class="card-box table-responsive">
                     <table id="datatable-checkbox" class="table table-striped table-bordered bulk_action" style="width:100%; white-space:nowrap;">
                       <thead>
                         <tr>
@@ -35,6 +35,8 @@
                           <th>탈퇴여부</th>
                           <th>활동지역</th>
                           <th>포인트총액</th>
+                          <th>구매자경험치</th>
+                          <th>판매자경험치</th>
                           <th>등급</th>
                           <th>가입날짜</th>
                           <th>업데이트날짜</th>
@@ -46,9 +48,8 @@
                        <tbody>
                      <tr th:if="${not #lists.isEmpty(userList)}" th:each="l:${userList}">
                      <td>
-                        <input type="checkbox" id="check-all" >
                      </td>
-                     	<td th:text="${l.userId}"></td>
+                     	<td th:text="${l.userIdCode}"></td>
                         <td th:text="${l.userPw}"></td>
 		                <td th:text="${l.userName}"></td>
 		                <td th:text="${l.userContact}"></td>
@@ -56,12 +57,14 @@
               			<td th:text="${l.userStatus}"></td>
               			<td th:text="${l.userArea}"></td>
               			<td th:text="${l.userPoint}"></td>
+              			<td th:text="${l.userRatingBuyer}"></td>
+              			<td th:text="${l.userRatingSeller}"></td>
               			<td th:text="${l.userRating}"></td>
               			<td th:text="${l.userRreDate}"></td>
               			<td th:text="${l.userUpDate}"></td>
               			<td th:text="${l.userLevel}"></td>
-              			<td></td>
-              			<td></td>
+              			<td><button type="button" class="btn btn-primary" style="font-size: 10px">수정</button></td>
+						<td><button type="button" class="btn btn-danger" style="font-size: 10px">삭제</button></td>
                         <!-- <td><button type="button" th:href="@{/pointList(pointDetailsCode=${l.pointDetailsCode})}" class="btn btn-danger" style="font-size: 10px">수정</button></td>
                         <td><button type="button" th:href="@{/pointList(pointDetailsCode=${l.pointDetailsCode})}" class="btn btn-danger" style="font-size: 10px">삭제</button></td> -->
                      </tr>
