+++ ks43team01/src/main/resources/templates/userpage/user/sellerjoin.html
@@ -87,7 +87,8 @@
 		<div class="container">
 			<div class="input-form-backgroud row">
 				<div class="input-form col-md-6 mx-auto">
-					<h4 class="mb-3" id="main">판매자추가가입</h4>
+					<h3 class="mb-3" id="main">나만의 기술스택을 팔아보아요!</h3>
+					<h4 class="mb-3" id="main">판매회원추가가입</h4>
 						<form class="validation-form" novalidate id="sellerBusiness" th:action="@{sellerBusiness}" method="post">
 							<div class="row">
 								<div class="col-md-6 mb-4">
