+++ ks43team01/src/main/resources/templates/userpage/user/userinfomation.html
@@ -116,7 +116,7 @@ layout:decorate="~{userpage/layout/userdefault}">
 								  		</div>
 								      	<div class="form-group">
 										    <label for="expertRetentionTechnology" class="col-12">보유기술</label>
-										    <input type="text" class="form-control col-6" id="expertRetentionTechnology">
+										    <input type="text" class="form-control col-6" id="expertRetentionTechnology" >
 								  		</div>
 								      	<div class="form-group">
 										    <label for="expertIntroduction" class="col-12">자기소개</label>
